Implement an automated birthday email for People's object
Added a Birthday field and a IsBirthdayEmailEnable flag to
the Person object. A daily background task checks for
matching birthdays and sends a predefined email if
enabled. For example, if \"John Doe\" has a birthday on
June 6th and the flag is active, the system will
automatically send a greeting email that day.
The flag can be toggled on and off on each person's record
by pressing its respective checkbox. Such, is only possible
if both email and birthday fields are provided by the user.
Existent hooks and functions were reused before commiting
to creating dedicated ones.
The checkbox componente used is a native html one, as
a placeholder for the custom buttons of the Twenty team.
The feature works, however it is only a simple suggestion
of the whole feature in itself.

Co-authored-by: Manuel Monteirinho Fernandes <manuel.monteirinho@tecnico.ulisboa.pt>